# backend/internal/handler/board_handler.go
package handler

import (
	"net/http"

	"github.com/aumputthipong/mini-erp-kanban/backend/internal/dto"
	"github.com/aumputthipong/mini-erp-kanban/backend/internal/httputil"
	"github.com/aumputthipong/mini-erp-kanban/backend/internal/mapper"
	"github.com/aumputthipong/mini-erp-kanban/backend/internal/middleware"
	"github.com/aumputthipong/mini-erp-kanban/backend/internal/service"
	"github.com/google/uuid"
)

type BoardHandler struct {
	boardService service.BoardServicer
}

func NewBoardHandler(boardService service.BoardServicer) *BoardHandler {
	return &BoardHandler{
		boardService: boardService,
	}
}

// GetAllBoards lists every board the caller is a member of.
//
// @Summary  List my boards
// @Tags     boards
// @Produce  json
// @Security CookieAuth
// @Success  200 {array}  dto.BoardSummaryResponse
// @Failure  401 {object} httputil.ErrorResponse
// @Router   /api/boards [get]
func (h *BoardHandler) GetAllBoards(w http.ResponseWriter, r *http.Request) error {
	userID, ok := r.Context().Value(middleware.UserIDKey).(string)
	if !ok || userID == "" {
		return httputil.NewAPIError(http.StatusUnauthorized, "Unauthorized", nil)
	}
	boards, err := h.boardService.GetAllBoards(r.Context(), userID)
	if err != nil {
		return httputil.NewAPIError(http.StatusInternalServerError, "Failed to fetch boards", err)
	}

	result := make([]dto.BoardSummaryResponse, 0, len(boards))
	for _, b := range boards {
		result = append(result, dto.BoardSummaryResponse{
			ID:         b.ID,
			Title:      b.Title,
			UpdatedAt:  b.UpdatedAt,
			TotalCards: b.TotalCards,
			DoneCards:  b.DoneCards,
			Members:    b.Members,
		})
	}
	httputil.RespondJSON(w, http.StatusOK, result)
	return nil
}

// GetBoardData returns the columns + cards for one board (full board view).
//
// @Summary  Get board contents
// @Tags     boards
// @Produce  json
// @Security CookieAuth
// @Param    boardID path     string  true  "Board UUID"
// @Success  200     {array}  dto.ColumnResponse
// @Failure  400     {object} httputil.ErrorResponse
// @Failure  404     {object} httputil.ErrorResponse "not a member or board not found"
// @Router   /api/boards/{boardID} [get]
func (h *BoardHandler) GetBoardData(w http.ResponseWriter, r *http.Request) error {
	boardID, err := httputil.GetUUIDParam(r, "boardID")
	if err != nil {
		return httputil.NewAPIError(http.StatusBadRequest, "Invalid board ID format", err)
	}

	columns, err := h.boardService.GetBoardWithCards(r.Context(), boardID)
	if err != nil {
		return httputil.NewAPIError(http.StatusInternalServerError, "Failed to fetch board data", err)
	}

	httputil.RespondJSON(w, http.StatusOK, mapper.ToColumnResponses(columns))
	return nil
}
// CreateBoard creates a new board owned by the caller.
//
// @Summary  Create board
// @Tags     boards
// @Accept   json
// @Produce  json
// @Security CookieAuth
// @Param    payload body     dto.CreateBoardRequest true  "Board title"
// @Success  201     {object} map[string]string      "id"
// @Failure  400     {object} httputil.ErrorResponse
// @Router   /api/boards [post]
func (h *BoardHandler) CreateBoard(w http.ResponseWriter, r *http.Request) error {
	var req dto.CreateBoardRequest
	if err := httputil.DecodeAndValidate(r, &req); err != nil {
		return err
	}

	userIDStr, ok := r.Context().Value(middleware.UserIDKey).(string)
	if !ok || userIDStr == "" {
		return httputil.NewAPIError(http.StatusUnauthorized, "Unauthorized", nil)
	}
	// validate UUID format ของ userID จาก token
	if _, err := uuid.Parse(userIDStr); err != nil {
		return httputil.NewAPIError(http.StatusInternalServerError, "Invalid user ID in token", err)
	}

	boardID, err := h.boardService.CreateBoard(r.Context(), req.Title, userIDStr)
	if err != nil {
		return httputil.NewAPIError(http.StatusInternalServerError, "Failed to create board", err)
	}

	httputil.RespondJSON(w, http.StatusCreated, map[string]string{"id": boardID})
	return nil
}

// UpdateBoard updates a board's title and/or budget. Manager+ only.
//
// @Summary  Update board
// @Tags     boards
// @Accept   json
// @Produce  json
// @Security CookieAuth
// @Param    boardID path     string                 true  "Board UUID"
// @Param    payload body     dto.UpdateBoardRequest true  "Fields to update"
// @Success  200     {object} dto.BoardSummaryResponse
// @Failure  400     {object} httputil.ErrorResponse
// @Failure  403     {object} httputil.ErrorResponse "not a manager"
// @Router   /api/boards/{boardID} [patch]
func (h *BoardHandler) UpdateBoard(w http.ResponseWriter, r *http.Request) error {
	boardID, err := httputil.GetUUIDParam(r, "boardID")
	if err != nil {
		return httputil.NewAPIError(http.StatusBadRequest, "Invalid or missing board ID", err)
	}

	var req dto.UpdateBoardRequest
	if err := httputil.DecodeAndValidate(r, &req); err != nil {
		return err
	}

	updatedBoard, err := h.boardService.UpdateBoard(r.Context(), boardID, req.Title, req.Budget)
	if err != nil {
		return httputil.NewAPIError(http.StatusInternalServerError, "Failed to update board", err)
	}

	httputil.RespondJSON(w, http.StatusOK, updatedBoard)
	return nil
}

// MoveToTrash soft-deletes a board (sets deleted_at). Owner only.
//
// @Summary  Move board to trash
// @Tags     boards
// @Security CookieAuth
// @Param    boardID path string true "Board UUID"
// @Success  204
// @Failure  403 {object} httputil.ErrorResponse "not the owner"
// @Router   /api/boards/{boardID} [delete]
func (h *BoardHandler) MoveToTrash(w http.ResponseWriter, r *http.Request) error {
	boardID, err := httputil.GetUUIDParam(r, "boardID")
	if err != nil {
		return httputil.NewAPIError(http.StatusBadRequest, "Invalid board ID format", err)
	}

	if err := h.boardService.MoveBoardToTrash(r.Context(), boardID); err != nil {
		return httputil.NewAPIError(http.StatusInternalServerError, "Failed to move board to trash", err)
	}

	w.WriteHeader(http.StatusNoContent)
	return nil
}

// GetTrash lists soft-deleted boards owned by the caller.
//
// @Summary  List trashed boards
// @Tags     trash
// @Produce  json
// @Security CookieAuth
// @Success  200 {array} dto.TrashedBoardDTO
// @Router   /api/trash [get]
func (h *BoardHandler) GetTrash(w http.ResponseWriter, r *http.Request) error {
	userID, ok := r.Context().Value(middleware.UserIDKey).(string)
	if !ok || userID == "" {
		return httputil.NewAPIError(http.StatusUnauthorized, "Unauthorized", nil)
	}
	boards, err := h.boardService.GetTrashedBoards(r.Context(), userID)
	if err != nil {
		return httputil.NewAPIError(http.StatusInternalServerError, "Failed to get trash", err)
	}

	httputil.RespondJSON(w, http.StatusOK, mapper.ToTrashedBoardDTOs(boards))
	return nil
}

// HardDelete permanently removes a trashed board. Owner only.
//
// @Summary  Permanently delete trashed board
// @Tags     trash
// @Security CookieAuth
// @Param    boardID path string true "Board UUID"
// @Success  204
// @Failure  403 {object} httputil.ErrorResponse
// @Router   /api/trash/{boardID} [delete]
func (h *BoardHandler) HardDelete(w http.ResponseWriter, r *http.Request) error {
	boardID, err := httputil.GetUUIDParam(r, "boardID")
	if err != nil {
		return httputil.NewAPIError(http.StatusBadRequest, "Invalid board ID format", err)
	}

	if err := h.boardService.HardDeleteBoard(r.Context(), boardID); err != nil {
		return httputil.NewAPIError(http.StatusInternalServerError, "Delete failed", err)
	}

	w.WriteHeader(http.StatusNoContent)
	return nil
}

// RestoreBoard restores a soft-deleted board. Owner only.
//
// @Summary  Restore trashed board
// @Tags     trash
// @Security CookieAuth
// @Param    boardID path string true "Board UUID"
// @Success  204
// @Router   /api/trash/{boardID}/restore [patch]
func (h *BoardHandler) RestoreBoard(w http.ResponseWriter, r *http.Request) error {
	boardID, err := httputil.GetUUIDParam(r, "boardID")
	if err != nil {
		return httputil.NewAPIError(http.StatusBadRequest, "Invalid board ID format", err)
	}

	if err := h.boardService.RestoreBoard(r.Context(), boardID); err != nil {
		return httputil.NewAPIError(http.StatusInternalServerError, "Failed to restore board", err)
	}

	w.WriteHeader(http.StatusNoContent)
	return nil
}